refactor: change contract types exports
